+++ addon/mixins/flexberry-objectlistview-on-edit-form-controller.js
@@ -4,18 +4,18 @@
 
 import Ember from 'ember';
 import { Query } from 'ember-flexberry-data';
+import PredicateFromFiltersMixin from '../mixins/predicate-from-filters';
 import deserializeSortingParam from '../utils/deserialize-sorting-param';
-
-const { Condition, SimplePredicate, StringPredicate, ComplexPredicate, DatePredicate, Builder } = Query;
 
 /**
   Mixin for edit-form-controller for ObjectListView support.
 
   @class FlexberryObjectlistviewOnEditFormControllerMixin
   @extends Ember.Mixin
+  @uses PredicateFromFiltersMixin
   @public
 */
-export default Ember.Mixin.create({
+export default Ember.Mixin.create(PredicateFromFiltersMixin, {
   /**
     Name of related to FOLV edit form route.
 
@@ -187,82 +187,6 @@ export default Ember.Mixin.create({
   }),
 
   /**
-    Return predicate to filter through.
-
-    @example
-      ```javascript
-      // app/routes/example.js
-      ...
-      predicateForFilter(filter) {
-        if (filter.type === 'string' && filter.condition === 'like') {
-          return new StringPredicate(filter.name).contains(filter.pattern);
-        }
-
-        return this._super(...arguments);
-      },
-      ...
-      ```
-
-    @method predicateForFilter
-    @param {Object} filter Object (`{ name, condition, pattern }`) with parameters for filter.
-    @return {BasePredicate|null} Predicate to filter through.
-  */
-  predicateForFilter(filter) {
-    if (filter.pattern && filter.condition) {
-      switch (filter.type) {
-        case 'string':
-          return filter.condition === 'like' ?
-            new StringPredicate(filter.name).contains(filter.pattern) :
-            new SimplePredicate(filter.name, filter.condition, filter.pattern);
-        case 'boolean':
-          return new SimplePredicate(filter.name, filter.condition, filter.pattern);
-        case 'number':
-          return new SimplePredicate(filter.name, filter.condition, filter.pattern ? Number(filter.pattern) : filter.pattern);
-        case 'date':
-          return new DatePredicate(filter.name, filter.condition, filter.pattern);
-
-        default:
-          return null;
-      }
-    } else {
-      if (!filter.condition && filter.type === 'string') {
-        Ember.set(filter, 'condition', 'like');
-        return new StringPredicate(filter.name).contains(filter.pattern);
-      }
-    }
-
-    return null;
-  },
-
-  /**
-    Return predicate for `QueryBuilder` or `undefined`.
-
-    @method _filtersPredicate
-    @return {BasePredicate|undefined} Predicate for `QueryBuilder` or `undefined`.
-    @private
-  */
-  _filtersPredicate() {
-    let filters = this.get('filters');
-    if (filters) {
-      let predicates = [];
-      for (let filter in filters) {
-        if (filters.hasOwnProperty(filter)) {
-          let predicate = this.predicateForFilter(filters[filter]);
-          if (predicate) {
-            predicates.push(predicate);
-          } else {
-            delete filters[filter];
-          }
-        }
-      }
-
-      return predicates.length ? predicates.length > 1 ? new ComplexPredicate(Condition.And, ...predicates) : predicates[0] : undefined;
-    }
-
-    return undefined;
-  },
-
-  /**
     It forms the limit predicate for FOLV loaded data on edit form.
 
     By default it returns `undefined`.
@@ -294,7 +218,7 @@ export default Ember.Mixin.create({
       let hierarchicalAttribute = this.get('hierarchicalAttribute');
       let modelName = this.get('folvModelName');
       let projectionName = this.get('folvProjection');
-      let builder = new Builder(this.store)
+      let builder = new Query.Builder(this.store)
         .from(modelName)
         .selectByProjection(projectionName)
         .where(hierarchicalAttribute, 'eq', id);